suite: add ability to insert custom environment variables via text files
Allows the user to create text files defining additional environment variables to be appended during build time per repo. Not a perfect implementation, as whenever the suite fails on configure/cmake/meson/rust, the user will have to exit twice due to the suite invoking a subshell to temporarily define the user's environment variables. I'll have to find a better method for implementing this later